# More efficient serialization of medium `BigInt` values